[threads] Move linux arch files and create symbolic link for sim arch.
